refactored filter logic
